Use token as current session, auth route hits session controller
After some experimenting with the API and a user PORO, I opted to
rely fully on the API for information, since the access token has a
very short timespan anyway. Consequently, the session now stores the token
which in the future I will create a helper method to give other models access to it.