Create user model with 23 info, add sessions controller
When a user authorizes 23 and Me, it looks for that user based on their 23 and Me id.
If none is found, it creates a new user, storing as well their access and refresh tokens.
Adds current user method. Gives message to user that they're logged in as their email.
Hides login/new user ("register kit") once logged in
Future refactoring will have to take into account their token expiring but moving on to outputting information next