Add profile controller and images, last name
Just need to output haplogroups to the view and the next step would then be
implementing edit profiles.

23AndMe returns URLs for profile images, but in the case where the user hasn't
set one, it returns a path to a default image that doesn't work. So to be good,
the code would need a conditional to check for that and set our own default image
instead.